--- Ejercicio4/nbproject/private/private.properties
@@ -1,1 +1,0 @@
-user.properties.file=/home/jesus/.netbeans/20/build.properties
