--- Ejercicio4/nbproject/private/private.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
-    <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/home/jesus/Escritorio/SEGUNDO%20CUATRI/DS/practica1/Ejercicio4/src/ejercicio4/CadenaFiltros.java</file>
-            <file>file:/home/jesus/Escritorio/SEGUNDO%20CUATRI/DS/practica1/Ejercicio4/src/ejercicio4/Cliente.java</file>
-            <file>file:/home/jesus/Escritorio/SEGUNDO%20CUATRI/DS/practica1/Ejercicio4/src/controlador/Mandos.java</file>
-            <file>file:/home/jesus/Escritorio/SEGUNDO%20CUATRI/DS/practica1/Ejercicio4/src/ejercicio4/GestorFiltros.java</file>
-            <file>file:/home/jesus/Escritorio/SEGUNDO%20CUATRI/DS/practica1/Ejercicio4/src/controlador/EstadoMotor.java</file>
-        </group>
-    </open-files>
-</project-private>
