--- Ejercicio4/src/controlador/EstadoMotor.java
@@ -1,16 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Enum.java to edit this template
- */
-package controlador;
-
-/**
- *
- * @author carme
- */
-public enum EstadoMotor {
-    APAGADO,
-    ENCENDIDO,
-    ACELERANDO,
-    FRENANDO;
-}
