--- Ejercicio4/src/ejercicio4/Filtro.java
@@ -1,13 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
- */
-package ejercicio4;
-
-/**
- *
- * @author carme
- */
-public interface Filtro {
-    
-}
