+++ practica1/Ejercicio4/src/controlador/Mandos.java
@@ -4,10 +4,58 @@
  */
 package controlador;
 
+import javax.swing.JLabel;
+import javax.swing.JToggleButton;
+
 /**
  *
- * @author carme
+ * @author jesus
  */
+
 public class Mandos {
+    private JToggleButton botonAcelerador;
+    private JToggleButton botonEncendido;
+    private JToggleButton botonFreno;
+    private JLabel estado;
+    private EstadoMotor comandoActual; 
+    private Cliente cliente; 
+    private Salpicadero salpicadero;
+
     
+    public Mandos(Cliente cliente, Salpicadero salpicadero) {
+        this.cliente = cliente;
+        this.salpicadero = salpicadero;
+    }
+
+    
+    public void run() {
+        
+        botonAcelerador.addActionListener(e -> Acelerar());
+        botonEncendido.addActionListener(e -> Encender());
+        botonFreno.addActionListener(e -> Frenar());
+    }
+
+    
+    public void Encender() {
+        comandoActual = EstadoMotor.ENCENDIDO;
+        estado.setText("Encendido");
+        cliente.peticionGestorFiltros(); 
+        salpicadero.actualizarEstado(comandoActual); 
+    }
+
+    
+    public void Acelerar() {
+        comandoActual = EstadoMotor.ACELERANDO;
+        estado.setText("Acelerando");
+        cliente.peticionGestorFiltros(); 
+        salpicadero.actualizarEstado(comandoActual); 
+    }
+
+    
+    public void Frenar() {
+        comandoActual = EstadoMotor.FRENANDO;
+        estado.setText("Frenando");
+        cliente.peticionGestorFiltros(); 
+        salpicadero.actualizarEstado(comandoActual); 
+    }
 }
