+++ practica1/Ejercicio4/src/ejercicio4/Cliente.java
@@ -6,8 +6,16 @@ package ejercicio4;
 
 /**
  *
- * @author carme
+ * @author jesus
  */
 public class Cliente {
+    private GestorFiltros gestorFiltros;
     
+    public Cliente(GestorFiltros gestor){
+        this.gestorFiltros = gestor;
+    }
+    
+    public void peticionGestorFiltros(){
+        gestorFiltros.peticionCadenaFiltros();
+    }
 }
