+++ practica1/memoria.tex
@@ -1,4 +1,56 @@
-\documentclass{article}
+%========================= Classe du Document ==========================
+\documentclass[a4paper,fleqn]{article}
+%============================= Packages ================================
+\usepackage[margin=2.5cm]{geometry}
+\usepackage[hidelinks]{hyperref}
+\usepackage[utf8]{inputenc}
+\usepackage[T1]{fontenc}
+\usepackage{graphicx}
+\usepackage{listings}
+\usepackage{amssymb}
+\usepackage{xcolor}
+\usepackage{minted}
+\usepackage{svg} 
+\setlength{\parindent}{0pt}
+%=======================================================================
+
+
+
+%========================= Debut du document ===========================
 \begin{document}
-Memoria
+%=======================================================================
+
+
+
+
+%^^^^^^^^^^^^^^^^^^^^^^^^^^^ Page de Garde ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+\definecolor{color}{HTML}{F5F5F5}
+\thispagestyle{empty}
+
+\begin{figure}
+    \centering
+    \includegraphics[width=8cm]{logo.png}
+\end{figure}
+
+\vspace{6cm}	
+
+\begin{center}	
+    {\huge \bf\ Titre du Document  \\
+    \vspace{4cm}
+    \large Matière - Semestre \\
+
+    \vspace{0.5cm}
+
+    Prénom Nom
+    \vfill	
+    Mois Année 
+    \pagebreak}
+\end{center}
+%^^^^^^^^^^^^^^^^^^^^^^^^^ Fin Page de Garde ^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+
+
+
+%========================== Fin du Document ============================
 \end{document}
+%=======================================================================
